+++ tests/fixtures/noqa.py
@@ -20,9 +20,10 @@ formatted_string = f'Hi, {full_name}'  # noqa: Z305
 def some():  # noqa: Z110
     from my_module import some_function  # noqa: Z435
 
-    class Nested(object): ...  # noqa: Z220, Z431
+    class Nested(object): ...  # noqa: Z220, Z431, Z444
 
-    def nested(): ...  # noqa: Z430
+    def nested():  # noqa: Z430
+        anti_z444 = 1
 
     raise NotImplemented  # noqa: Z423
 
@@ -45,7 +46,7 @@ def many_locals():  # noqa: Z210
 
 
 def many_arguments(_arg1, _arg2, _arg3, _arg4, _arg5, _arg6):  # noqa: Z211
-    ...
+    anti_z444 = 1
 
 
 def many_returns(xy):  # noqa: Z212
@@ -79,23 +80,23 @@ def many_expressions(xy):  # noqa: Z213
 
 
 class ManyParents(dict, list, tuple, Exception):  # noqa: Z215
-    ...
+    anti_z444 = 1
 
 
 line = some.call(7 * 2, 3 / 4) / some.run(5 / some, 8 - 2 + 1)  # noqa: Z221
 if line and line > 2 and line > 3 and line > 4 and line > 5:  # noqa: Z221,Z222
-    ...
+    anti_z444 = 1
 
 if line:  # noqa: Z223
-    ...
+    anti_z444 = 1
 elif line > 1:
-    ...
+    anti_z444 = 1
 elif line > 2:
-    ...
+    anti_z444 = 1
 elif line > 3:
-    ...
+    anti_z444 = 1
 elif line > 4:
-    ...
+    anti_z444 = 1
 
 
 numbers = [
@@ -112,17 +113,17 @@ class BadClass:  # noqa: Z306
 
     @staticmethod  # noqa: Z433
     def some_static():
-        ...
+        anti_z444 = 1
 
     @staticmethod  # noqa: Z433
     async def some_async_static():
-        ...
+        anti_z444 = 1
 
     def __del__(self, *_args, **_kwargs):  # noqa: Z434
-        ...
+        anti_z444 = 1
 
     class Nested:  # noqa: Z306,Z431
-        ...
+        anti_z444 = 1
 
 
 magic_numbers = 13.2 + 50  # noqa: Z432
@@ -138,38 +139,38 @@ binary_number = 0B1001  # noqa: Z310
 number_with_scientific_notation = 1.5E+10  # noqa: Z310
 
 if '6' in nodes in '6':  # noqa: Z311
-    ...
+    anti_z444 = 1
 
 assert hex_number == hex_number  # noqa: Z312
 
 
-def test_function():
+async def test_function():
     return(123, 33)  # noqa: Z313
 
 
 if True:  # noqa: Z314
-    ...
+    anti_z444 = 1
 
 
 class SomeClass(FirstParent, SecondParent, object):  # noqa: Z315
-    ...
+    anti_z444 = 1
 
 
 try:
-    ...
+    anti_z444 = 1
 except BaseException:  # noqa: Z424
-    ...
+    anti_z444 = 1
 
 for symbol in 'abc':  # noqa: Z436
-    ...
+    anti_z444 = 1
 else:
-    ...
+    anti_z444 = 1
 
 
 try:  # noqa: Z437
-    ...
+    anti_z444 = 1
 finally:
-    ...
+    anti_z444 = 1
 
 nodes = nodes  # noqa: Z438
 
@@ -183,6 +184,10 @@ for index in range(6):  # noqa: Z442
     print(lambda: index)
 
 
-def function_with_unreachable():
+async def function_with_unreachable():
+    await test_function()
     raise ValueError()
     print(1)  # noqa: Z443
+
+
+1 + 2  # noqa: Z444