+++ tests/test_checker/test_noqa.py
@@ -103,6 +103,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z441': 1,
         'Z442': 1,
         'Z443': 1,
+        'Z444': 2,
     }
 
     process = subprocess.Popen(